- fixes #2120
